+++ src/test/java/service/LottoNumberMatcherTest.java
@@ -1,4 +1,6 @@
-import domain.LottoNumberMatcher;
+package service;
+
+import service.LottoNumberMatcher;
 import domain.lotto.LottoNumber;
 import domain.lotto.LottoPrize;
 import domain.lotto.LottoTicket;
