+++ src/test/java/service/LottoPrizeCounterTest.java
@@ -1,9 +1,12 @@
-import domain.LottoPrizeCounter;
+package service;
+
+import service.LottoPrizeCounter;
 import domain.lotto.*;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import service.LottoPrizeResult;
 
 import java.util.Arrays;
 import java.util.HashMap;
