+++ src/test/java/service/LottoPrizeResultTest.java
@@ -1,9 +1,13 @@
+package service;
+
 import domain.lotto.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import service.LottoPrizeResult;
+import service.LottoShop;
 
 import java.util.HashMap;
